+++ src/yoyo/Exceptions/ComponentMethodNotFound.php
@@ -7,7 +7,7 @@ class ComponentMethodNotFound extends \Exception
     public function __construct($component, $method)
     {
         parent::__construct(
-            "Unable to call component method. Public method [{$method}] not found on component: [{$component}]"
+            "Public method [{$method}] not found on Yoyo component [{$component}]"
         );
     }
 }