+++ src/yoyo/Exceptions/NonPublicComponentMethodCall.php
@@ -6,6 +6,6 @@ class NonPublicComponentMethodCall extends \Exception
 {
     public function __construct($componentName, $method)
     {
-        parent::__construct("[$componentName] component method [$method] not found.");
+        parent::__construct("Unable to call non-public method [$method] in Yoyo component [$componentName].");
     }
 }