+++ src/yoyo/QueryString.php
@@ -44,7 +44,9 @@ class QueryString
      */
     public function getPageQueryParams()
     {
-        $currentUrl = $this->currentUrl;
+        if (! $this->currentUrl) {
+            return [];
+        }
 
         // Filter out keys that are not explicitly set in the component queryString property
 
@@ -52,19 +54,19 @@ class QueryString
 
         // Get current query string values and merge them with new ones
 
-        parse_str(parse_url(htmlspecialchars_decode($currentUrl), PHP_URL_QUERY), $args);
+        $queryString = parse_url(htmlspecialchars_decode($this->currentUrl), PHP_URL_QUERY) ?? '';
+
+        parse_str($queryString, $args);
 
         $queryParams = array_merge($args, $new);
 
         // If a query string value matches the default value, remove it from the URL
 
         foreach ($queryParams as $key => $val) {
-            if (isset($this->defaults[$key]) && $val == $this->defaults[$key]) {
+            if (isset($this->defaults[$key]) && $val === $this->defaults[$key] || $val === '') {
                 unset($queryParams[$key]);
             }
         }
-
-        $queryParams = array_filter($queryParams);
 
         return $queryParams;
     }