+++ src/yoyo/Services/Configuration.php
@@ -10,14 +10,43 @@ class Configuration
 
     private static $options;
 
-    public static $htmx = '1.3.1';
+    public static $htmx = '1.8.4';
 
+    protected static $allowedConfigOptions = [
+        'addedClass',
+        'allowEval',
+        'attributesToSettle',
+        'defaultFocusScroll',
+        'defaultSettleDelay',
+        'defaultSwapDelay',
+        'defaultSwapStyle',
+        'disableSelector',
+        'historyCacheSize',
+        'historyEnabled',
+        'includeIndicatorStyles',
+        'indicatorClass',
+        'inlineScriptNonce',
+        'refreshOnHistoryMiss',
+        'requestClass',
+        'scrollBehavior',
+        'settlingClass',
+        'swappingClass',
+        'timeout',
+        'useTemplateFragments',
+        'withCredentials',
+        'wsReconnectDelay',
+    ];
+        
     public function __construct($options)
     {
         self::$options = array_merge([
             'namespace' => 'App\\Yoyo\\',
-            'defaultSwap' => 'outerHTML',
+            'defaultSwapStyle' => 'outerHTML',
             'historyEnabled' => false,
+            'indicatorClass' => 'yoyo-indicator',
+            'requestClass' => 'yoyo-request',
+            'settlingClass' => 'yoyo-settling',
+            'swappingClass' => 'yoyo-swapping',
         ], $options);
     }
 
@@ -39,7 +68,7 @@ class Configuration
     public static function htmxSrc(): string
     {
         if (empty($htmxSrc = self::get('htmx'))) {
-            $htmxSrc = 'https://unpkg.com/htmx.org@'.self::$htmx.'/dist/htmx.js';
+            $htmxSrc = 'https://unpkg.com/htmx.org@'.self::$htmx.'/dist/htmx.min.js';
         }
 
         return $htmxSrc;
@@ -66,19 +95,12 @@ HTML;
     public static function javascriptInitCode($includeScriptTag = true): string
     {
         $yoyoRoute = self::get('url', '');
-        $defaultSwap = self::get('defaultSwap', 'outerHTML');
-        $historyEnabled = self::get('historyEnabled', false) ? 'true' : 'false';
+        $configuration = array_intersect_key(static::$options, array_flip(static::$allowedConfigOptions));
+        $yoyoConfig = json_encode($configuration);
 
         $script = <<<HTML
         Yoyo.url = '{$yoyoRoute}';
-        Yoyo.config({
-            defaultSwapStyle: '{$defaultSwap}',
-            historyEnabled: {$historyEnabled},
-            indicatorClass:	'yoyo-indicator',
-            requestClass:	'yoyo-request',
-            settlingClass:	'yoyo-settling',
-            swappingClass:	'yoyo-swapping'
-        });
+        Yoyo.config($yoyoConfig);
 HTML;
 
         if ($includeScriptTag) {