+++ src/yoyo/YoyoCompiler.php
@@ -27,17 +27,6 @@ class YoyoCompiler
 
     protected $idCounter = 1;
 
-    /**
-     * These will automatically receive a `method` attribute.
-     */
-    protected $reactiveTags = [
-        'a',
-        'button',
-        'input',
-        'select',
-        'textarea',
-    ];
-
     public const HTMX_REQUEST_METHOD_ATTRIBUTES = [
         'boost',
         'delete',
@@ -50,8 +39,9 @@ class YoyoCompiler
     ];
 
     public const YOYO_ATTRIBUTES = [
-        // 'boost', // Has a conflict with component history cache
         'confirm',
+        'disable',
+        'disinherit',
         'encoding',
         'ext',
         'headers',
@@ -63,10 +53,16 @@ class YoyoCompiler
         'preserve',
         'prompt',
         'push-url',
+        'replace-url',
+        'request',
+        'select-oob',
         'select',
         'swap-oob',
         'swap',
+        'sync',
         'target',
+        'trigger',
+        'validate',
         'vals',
     ];
 
@@ -437,10 +433,9 @@ class YoyoCompiler
             }
         }
 
-        // Automatically make element reactive if it has the yoyo attribute, or if it's a clickable element
-        if ($element->hasAttribute(self::YOYO_PREFIX) || in_array($element->tagName, $this->reactiveTags)) {
+        // Automatically add the default hx-get="render" request to component root nodes and any child with the `yoyo` attribute
+        if ($element->hasAttribute(self::YOYO_PREFIX)) {
             $element->setAttribute(self::hxprefix('get'), self::COMPONENT_DEFAULT_ACTION);
-
             if (! $isRootNode) {
                 // Ensure re-active tags have an ID to improve swapping
                 $this->checkForIdAttribute($element);